requirements.txt
# SEO Toolkit — Python Dependencies

# HTTP requests
requests>=2.31.0
httpx>=0.27.0

# HTML parsing and crawling
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Data processing
pandas>=2.2.0

# Schema validation
jsonschema>=4.21.0

# Readability analysis
textstat>=0.7.3

# Image analysis
Pillow>=10.2.0

# XML sitemap parsing
defusedxml>=0.7.1

# Environment variables
python-dotenv>=1.0.0

# DataForSEO client (if they have one, otherwise raw HTTP)
# dataforseo-client

# Google API (for Search Console)
google-api-python-client>=2.100.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0

# Report generation
jinja2>=3.1.0

# Data models and validation
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Property-based testing
hypothesis>=6.98.0

# HTTP mocking for tests
pytest-httpx>=0.30.0
